% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_article_text.R
\name{get_article_text}
\alias{get_article_text}
\title{Get article text}
\usage{
get_article_text(url, timeout = NULL)
}
\arguments{
\item{url}{a single character string or character vector.}

\item{timeout}{}
}
\description{
Get article text
}
\examples{
get_article_text(url = "http://www.zeit.de/kultur/film/2018-04/tatort-frankfurt-unter-kriegern-obduktionsbericht")

}
