% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_client.R
\name{get_client}
\alias{get_client}
\title{Client status and API usage}
\usage{
get_client(api_key = Sys.getenv("ZEIT_ONLINE_KEY"))
}
\arguments{
\item{api_key}{character. The personal api code. To request an API key see: \url{http://developer.zeit.de/quickstart/} This parameter is by default set to the R Environment.}
}
\value{
a list of information about the client and API usage
}
\description{
\code{get_cleint} returns API access status and usage.
}
\examples{
\dontrun{
get_client()
}

}
\author{
Jan Dix (\email{jan.dix@uni-konstanz.de})
}
