% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content.R
\name{get_content}
\alias{get_content}
\title{Query content endpoint}
\usage{
get_content(query, limit = 10, offset = 0, sort = "release_date asc",
  begin_date = NULL, end_date = NULL,
  api_key = Sys.getenv("ZEIT_ONLINE_KEY"))
}
\arguments{
\item{api_key}{}
}
\description{
Query content endpoint
}
\examples{
get_content(query = "Merkel")

}
