% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_content_all.R
\name{get_content_all}
\alias{get_content_all}
\title{Query content endpoint}
\usage{
get_content_all(query, timeout = 2, begin_date = NULL, end_date = NULL,
  api_key = Sys.getenv("ZEIT_ONLINE_KEY"))
}
\arguments{
\item{api_key}{}
}
\description{
Query content endpoint
}
\examples{
get_content_all(query = "Merkel")

}
