% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzeit2.R
\docType{package}
\name{rzeit2}
\alias{rzeit2}
\alias{rzeit2-package}
\title{Client for the ZEIT ONLINE Content API}
\description{
Interface to gather newspaper articles from DIE ZEIT and ZEIT ONLINE,
based on a multilevel query. A personal API key is required for usage.
}
\references{
\url{http://developer.zeit.de}
}
\seealso{
\code{\link{get_content}} to expose a search in the ZEIT online archive,
\code{\link{get_content_all}} to get all results using \code{get_content},
\code{\link{get_client}} to get client information
}
