% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set api key to the .Renviron}
\usage{
set_api_key(api_key, path = stop("Please specify a path."))
}
\arguments{
\item{api_key}{character. The personal api code. To request an API key see: \url{http://developer.zeit.de/quickstart/}}

\item{path}{character. Path where the environment is stored. Default is the normalized path.}
}
\value{
None.
}
\description{
Function to set you API Key to the R environment when starting using \code{rzeit} package. Attention: You should only execute this functions once.
}
\examples{
\dontrun{
# this is not an actual api key
api_key <- "5t5yno5qqkufxis5q2vzx26vxq2hqej9"
set_api_key(api_key, tempdir())
}
}
\author{
Jan Dix <\email{jan.dix@uni-konstanz.de}>
}
